Change the way to handle bad request
This is mostly due to the fact that some web server doesn't handle the HEAD request. That said, graby will return an error in that case instead of trying using a GET request.

It mostly happend when requesting an asset url (image, pdf, etc ..). Instead of getting the content, we just get an empty doc.