If request fail try to fetch content anyway
Some content fail to be requested (because they return a code 50x or 40x) but they content a response.
Some website doesn’t work as expected and return an error when they actually send the good content.

So if the exception contain a response, we’ll now try to retrieve content from it