Full debug mode in Graby
This is an attempt to have a useful debugger in graby, in order to
identify quickly why a site config don't work or why some stuff are
missing in the body.

A big improvement would be export of the various steps in various files,
or just store the diff between.

---

Two levels of logging for graby

Graby can now be called with two different `log_level` in its config:
'info' or 'debug'.

'info' is similar to the previous debug, with information on steps
taken, without too much results in it.
'debug' is a full debug mode, with the HTML code/DOM stored at each
major steps of the fetching of the content, in a separate one-time log
file.

---

Two-level debug messages with backward compatibility

By default, 'debug' = false.
When using 'debug' = true, Logger::INFO messages are catched and stored in logs/graby.log.

Furthermore, it is possible to specify 'log_level' = 'debug' to store HTML code at different time of the process in logs/html.log (one-time file).