ContentExtractor: remove logic for wp img lazy loading
I triggered a bug on a page with a wordpress image lazy loading system
based on noscript tags. In the case of this page, two noscript tags were
nested for the same image. The page fetch returned two different results
depending on the libxml version installed on the Debian and Alpine
systems I used for the test.

As far as we already have a logic to move data-lazy-* values in src and
srcset, I consider that the whole logic around the noscript tag is not
needed anymore.

Signed-off-by: Kevin Decherf <kevin@kdecherf.com>