Add ability to process prefetched content
fetchContent() now accepts an optional parameter, prefetchedContent, which
can contain the content of a page that was fetched before calling Graby.

If we take the example of Wallabag it gives the ability of sending the
content of a page (through a browser extension for example) without
making network calls to fetch the page.

Signed-off-by: Kevin Decherf <kevin@kdecherf.com>