ContentExtractor: Fix encoding in JSON-LD
When page has an article body specified in JSON-LD, the HTML fragment is passed to `DOMDocument::loadHTML`.
`DOMDocument` will parse a HTML document as ISO-8859-1, unless the document contains an XML encoding declaration or HTML meta tag setting character set.

https://stackoverflow.com/a/39148511/160386

JSON-LD typically does not contain `meta[charset]` tag so we need to add it ourselves.

At this point the input should already be in UTF-8, as we convert all inputs to that encoding everywhere except in `Graby::cleanupHtml`. There we will clarify the requirement in doc comment.